+++ src/main/java/au/edu/rmit/sept/webapp/repository/KeywordRepository.java
@@ -3,6 +3,6 @@ package au.edu.rmit.sept.webapp.repository;
 import au.edu.rmit.sept.webapp.model.Keyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KeywordRepository extends JpaRepository<Keyword, Integer> {
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
     Keyword findByKeyword(String name);
 }
